+++ packages/components/CHANGELOG.md
@@ -14,6 +14,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ### [Unreleased]
 
+#### Changed
+
+- `Select`, `Multi Select`:
+  - gets rendered on `#top-layer` which enables it to be shown correctly even when used e.g. within a scroll container
+    ([#3754](https://github.com/porsche-design-system/porsche-design-system/pull/3754))
+  - gets positioned by [CSS Anchor Positioning](https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_anchor_positioning)
+    and [Floating UI](https://floating-ui.com/) as fallback for browsers not supporting it yet
+    ([#3754](https://github.com/porsche-design-system/porsche-design-system/pull/3754))
+  - modernized visual appearance of option list
+    ([#3754](https://github.com/porsche-design-system/porsche-design-system/pull/3754))
+
 ### [3.27.0-rc.0] - 2025-02-13
 
 #### Added